+++ playground/src/main/kotlin/team/duckie/quackquack/playground/realworld/TabPlayground.kt
@@ -58,4 +58,3 @@ fun QuackSubTabDemo() {
         },
     )
 }
-
