+++ ui-components/src/main/kotlin/team/duckie/quackquack/ui/component/divider.kt
@@ -7,6 +7,7 @@
 
 package team.duckie.quackquack.ui.component
 
+import android.annotation.SuppressLint
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -27,6 +28,7 @@ private object QuackDividerDefaults {
 /**
  * 덕키에서 사용되는 구분선(divider)을 그립니다.
  */
+@SuppressLint("KDocFields")
 @Composable
 public fun QuackDivider() {
     with(
