+++ ui-components/src/main/kotlin/team/duckie/quackquack/ui/component/tag.kt
@@ -31,6 +31,7 @@ import androidx.compose.ui.util.fastForEachIndexed
 import kotlinx.collections.immutable.ImmutableList
 import team.duckie.quackquack.ui.border.QuackBorder
 import team.duckie.quackquack.ui.color.QuackColor
+import team.duckie.quackquack.ui.component.internal.QuackSurface
 import team.duckie.quackquack.ui.component.internal.QuackText
 import team.duckie.quackquack.ui.icon.QuackIcon
 import team.duckie.quackquack.ui.textstyle.QuackTextStyle
@@ -492,7 +493,7 @@ public fun QuackRoundTag(
     isSelected: Boolean,
     onClick: (() -> Unit)? = null,
 ): Unit = QuackRoundTagInternal(
-    modifier = Modifier,
+    modifier = modifier,
     text = text,
     isSelected = isSelected,
     onClick = onClick,
